stop deleting ProductRelationType if used

+++ acondbs/schema/product_relation_type.py
@@ -2,6 +2,7 @@ import graphene
 from graphene_sqlalchemy import SQLAlchemyObjectType
 
 from ..models import Product as ProductModel
+from ..models import ProductRelation as ProductRelationModel
 from ..models import ProductRelationType as ProductRelationTypeModel
 
 from ..db.sa import sa
@@ -41,9 +42,11 @@ class DeleteProductRelationType(graphene.Mutation):
 
     def mutate(root, info, type_id):
         type_ = ProductRelationTypeModel.query.filter_by(type_id=type_id).first()
-        # products = ProductModel.query.filter_by(product_relation_type_id=product_relation_type_id).all()
-        # if products:
-        #     raise ValueError('Cannot delete the product type "{}". Products of this type exist'.format(product_relation_type.name))
+        products = ProductModel.query.join(
+            ProductRelationModel,
+            (ProductModel.product_id == ProductRelationModel.self_product_id)).join(ProductRelationTypeModel).filter(ProductRelationTypeModel.type_id==type_id).all()
+        if products:
+            raise ValueError('Cannot delete the product relation type "{}". Products with this relation type exist'.format(type_.name))
         sa.session.delete(type_)
         sa.session.commit()
         ok = True
